requirements.txt
mypy
pytest
pylint
grpcio
grpcio-tools